added unit test round 1
