+++ src/solid-forms.ts
@@ -1,3 +1,3 @@
 import { FormModel } from './classes/form-model'
-import { ShexFormModel } from './classes/shex-ui'
+import { ShexFormModel } from './classes/shex-form-model'
 export { FormModel, ShexFormModel }
