+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 import * as solidFetch from './solid-fetch'
-import * as rdfSolid from './rdf'
 
-export { solidFetch, rdfSolid }
+export { solidFetch }
