+++ test/__mocks__/n3.js
@@ -1,0 +1,19 @@
+const DataFactory = jest.fn()
+const Lexer = jest.fn()
+const Parser = jest.fn()
+const Writer = jest.fn()
+const Store = jest.fn()
+const StreamParser = jest.fn()
+const StreamWriter = jest.fn()
+const Util = jest.fn()
+
+module.exports = {
+  DataFactory,
+  Lexer,
+  Parser,
+  Writer,
+  Store,
+  StreamParser,
+  StreamWriter,
+  Util
+}
