--- test/solid-forms.test.ts
@@ -1,14 +1,0 @@
-import { FormModel } from '../src/classes/form-model'
-
-/**
- * Dummy test
- */
-describe('Dummy test', () => {
-  it('works if true is truthy', () => {
-    expect(true).toBeTruthy()
-  })
-
-  it('DummyClass is instantiable', () => {
-    expect(new FormModel('url path')).toBeInstanceOf(FormModel)
-  })
-})
